Made Schemas and some routes. With the help of Express, we made some routes in the app. First we created two new folder named "routes" and "modules". In modules folder we created two schemas ie. UseSchema and NoteSchema in User.js and Note.js file. We created schemas from websitea and after that exported those schemas as models of mongoose model.exports = mongoose.model('use', UserSchema)
In routes folder we created routes (here 2 routes) with help of express and express.Router()
In that router we used router.get('/', (callback)) and in callback we used req and res as argument and we return res.json() file.
We also exports that router using "module.exports = router"

Finally in index.js, we added those routes using app.use()
In  app.use, first parameter for what path to display content
and second parameter for where the js file of that route is store. ( used inside "require")

Affter command "npx nodemon index.js" the server starts and
in port 3000 when you pase the new routes, you can see the response their.