Connected to database and stored data. First we used proper mongoURI bacause of higher node.js version. After that use that schema format which we want to send data( Here we want to send User schema in auth.js). For sending data from page body to database on mongoDB we used get/post method of express.router. The post method is best to sending sensitive data on database. Used async await for saving data on database. use save() method to save data on database.
We also chenged the database connection code.
in connection we used thsen and catch methods.
then method is used for showing database connected starting
and catch is for if error occurred in connection.
So that we get satisfaction if database connected or error found while connecting.

For sending through we temporary use the thunder client (vs code feature).
In thunder client we write content in body section to send req.body of page.
Before that we used to add header as content-type to application/json
And for using json file type in react app we must use a middleware express.json() in index.js
like code "app.use(express.json())"
This code help to read json file type from body.