Added validation in entries using express.js First we install express-validator using "npm install express-validator" in terminal. Then we import express validator where we want to set validation by code "const { body, validationResult } = require("express-validator");" at top of page. Then in post() method after path string we set required validation in an array like "[     body("name", "Enter proper name").isLength({ min: 4 }),     body("email", "Enter valid email").isEmail(),     body("password", "Minimum length is 8 for password").isLength({ min: 8 }),   ]," Then in callback function of post method ie. (req,res) we first check the validation have error or not by code "const result = validationResult(req);" if it has no error then the variable result is empty Then we send that data to database by using code "    if (result.isEmpty()) {       const user = new User(req.body);       await user         .save()         .catch((err) => console.log(err))         .then(console.log("user is saved on database"));       return res.send(req.body);     }" And if there are some errors in validation then result is not empty then we print those errors like code "   result.array().forEach((e) => console.log(e.msg));     res.status(400).send({ errors: result.array() });"
This is how we set validation.
Also to set unique validation we made indexes in schema model so where we set unique property of schema as true it will not allow for dublicate entry
and produce error in dublicate entry, so we must catch that error (Here we catch error in auth.js while User.save())
So it will not affect app working.

Note: To set a new database folder from code then in dabase uri, use the folder name after (port number)/ and before first ? sign.
It will automatically create a folder in database when first time the query gets hit.