Getting user detatail from token. For this purpose we create a middleware ( a function which do job for us). we created a middleware named folder and in that folder we created a file named fetchUser.js In that file, we first imported jwt and our secret key for jwt operations, then we created arrow function of name fetchUser who takes three argument (req, res, next), In that function we first got token value from header of request named auth-token if no such token then send error else in try block first check that token matches to jwt secret key using verify method of jwt. if matched then store id from that object to req.user and after that call next() so that the function next to calling middleware is automatically called.
In auth.js
we created new route for path /api/auth/getUser in which that fetchUser middleware is called.
From the middleware we got user id,
using that id we find the user form User database which matches {_id : userId} using findOne nethod
and after that we select information from that user
and that information is send to response.

To check this process,
we created new thunderclient request in iNotebook collection which is named get user from token
then we set path to localhost:5000/api/auth/getUser
after that in header we create a new header name auth-token and
set value to token which we want to search for
after click on send we will get user information in the response if token is correct
else we will get error.
(we will get token from thuderclient request named "login a user")