Adding note and fetching all notes. We create two routes in note.js file, one for fetching all notes and other for saving a note in user database. For both the process we use the middleware named fetchuser which gives us the user id from token because these both routes works only when user is logged in. So for both the route, we send auth-token in header with token of user.
For fetching all notes, we simply used middleware fetchUser which will give us user id and
according to that user id we fetch all notes present in that user's database.
For this we use find() method to find user id.

For creating a note, we used post method of route,
also here after using fetchuser middleware, we used the validation so that no one can save empty note in database.
When saving the note, in note we save a user id send by fetchUser middleware as user: req.user and title, description and tag.
Then saving that note in database.
Also for creating a note, we must create a foreigh key in Note module which connected to user id of User module,
so we created user entry in Note schema and gave it type of mongoose.Schema.Types.ObjectId and also gave ref : "user" to it.
So it acts like foreign key and connect User schema with Note schema.

For testing, we created new folder in thunderclient's iNotebook with name note
in note folder,
we created first get request as fetch all notes and set path to  /api/note/fetchAllNotes
and also in header setting content-type to application/json
and auth-token as token
Also,
we created a post request as add a note and set path to /api/note/addNote
and setting header as content-type as application/json and auth-token as token