Adding note and fetching all notes. We create two routes in note.js file, one for fetching all notes and other to add note to database. In both the process, we use the middleware named fetchuser which gives us the user id from token, because these both routes works only when user is logged in. So for both the route, we send auth-token in header with token of user.
First route for fetching all notes of user.
We use get method abd set path to /fetchAllNotes.
For this we simply used the user id which we will get from fetchUser middleware
And with help of that user id we fetch all notes of user.
For that we used find method where {user: req.user}

Second route for adding a note in database.
For this we used post method.
We first made changes in Note schema of Note.js file,
we created an entry named user and gave type as mongoose.Schema.Types.ObjectId and ref as user.
So it will act as a foreign key in Note schema and which connect User schema with Note schema.
After that in note.js,
we used post method and set path to /addNote
after that we used fetchUser middleware and after that
we used the validation part so that no one can save empty note in databaser.
After that we created a new Note with user: req.user and title, description and tag entries.
We saved user id sent by fetchUser middleware to user entry in note so it related to a specific user.
After we saved the note and send note to response.

For testing, in thunderclient,
we created a new folder in iNotebook named fetchAllNotes,
in that we created a new get request for fetching all notes with path /api/note/fetchAllNotes
and also crested a second new post request for add a note with path /api/note/addNote
For both request, in header, we set content-type to application/json and auth-token to token of a use.
For add a note request, in body part, we must give a json file of note with title, description and tag in json format.