New route for updating a note. For updating a note from database, we use the PUT method. In put method, first argument is path, second is middleware which gave us user id from token, and third argument is (req,res). In path we send the id of note which we want to update. After that, in trycatch box we check the id from path is exist or not using findById method of schema. If not then send file no found error 404. After that, we check that the user id we get from middleware fetchUser and user id of our note (get from note id provided in path) are same or not. (the user id of note is in object form we have convert it into string before checking) If user id is different for both user and note then send unauthorize error 401. After that we destructure the data get from req.body. After that we used findIdAndUpdate method to make update in database.
For testing,
we make a new thunderClient PUT request in note folder of iNotebook.
And gave path like localhost:5000/api/note/updateNote/:id, where :id is id of note which we want to change.
value for :id is get from database of note, which is a note id not user id.
Set header, content-type and auth-token to request.
Then in body, we made a json file of updation in note
And then save and send it