Context api We want to create context api, so that any of the react component can easily access the state and if last component is using that state then we need not to send that state to each component only last component can directly use state using context api. Firstly we created a folder named context in src folder. Then in contect, we created note folder. In note folder, we created two files, noteContext.js and NoteState.js In noteContext.js, we created a context named noteContext using createContext from react. After in NoteState.js, First created a function with props argument and then using useState we created a state, and we return props.children in NoteContext.Provider and in NoteContext.Provider we send state as value.
After that in App.js where all components are listed.
Enclose the Router component inside that NOteState component, so all the component can access the state.

Like in About component,
first import useContext and useEffect and noteContext,
then define a variable with noteContext using useContext.
The useEffect is used if we created a function inside noteContext, so we can use using useEffect.
When we want to use state value from noteContext in About component, that time we used variableName.state.valueName

By this wey, we can directly use state in About component and not needed to pass state as props to previous of About component.