Use of fontawesome in code for icon use. The first step to use fontawesome kit in index.html file by one line js code in head section : <script src="https://kit.fontawesome.com/5e84f764cb.js" crossorigin="anonymous"></script>
Then in body we can use the one line code of icons we want to use.
Like here we used the delete icon and edit icon
<i className="fa-regular fa-trash-can mx-2"></i> {/* Fontawesome for delete button */}
<i className="fa-regular fa-pen-to-square mx-2"></i> {/* Fontawesome for edit button */}

We also add style to i tag in index.css to set cursor to pointer and increase the size of icon when over on them.