Connected API call First we used the connection between browser and api by using cors middleware in express For that in terminal of backend folder, to install cors we write: npm install cors and after installation in backend folder, in index.js of backend we set that cors middleware in use like :-, var express = require('express') var cors = require('cors') var app = express() app.use(cors())
After this we can call api from the browser side too and for that we run command "npm run both" in iNotebook app to run server and browser at same time.
After this, we set the browser must show the notes from database when browser loads.
For this we made a function getNotes in NoteState.js of iNotebook app's context folder
We used fetch method like:-
// Default options are marked with *
  const response = await fetch(url, {
    method: "POST", // *GET, POST, PUT, DELETE, etc.
//    mode: "cors", // no-cors, *cors, same-origin
//    cache: "no-cache", // *default, no-cache, reload, force-cache, only-if-cached
//    credentials: "same-origin", // include, *same-origin, omit
    headers: {
      "Content-Type": "application/json",
      // 'Content-Type': 'application/x-www-form-urlencoded',
    },
//    redirect: "follow", // manual, *follow, error
//    referrerPolicy: "no-referrer", // no-referrer, *no-referrer-when-downgrade, origin, origin-when-cross-origin, same-origin, strict-origin, strict-origin-when-cross-origin, unsafe-url
    body: JSON.stringify(data), // body data type must match "Content-Type" header
  });
  return response.json(); // parses JSON response into native JavaScript objects
}

by removing unneccessary part of above code.
Also for getNotes function we used try catch block to handle errors
That getNote function is also exported from NoteState.js
And we call this getNote function from Note.js of components folder.
for calling getNotes context we used useEffect tool so when browser page loads it shows the all notes from database to the browser.
Also since we used useEffect for calling the function, we use useCallback to memorize the function between re-renders.
For changes in above code for getNotes function,we used
method : "GET"
in header, we also send auth-token key whose value we will get from current users login.
We do not send body here, as it not needed when fetchin the notes from database through api.
After that we checked if response is ok or not , if not then throw an error.
Then we use a variable named json to store the response in json format (response.json) asynchroniously
And then set that variable to setNotes state, so that now notes state is changed.

We also make changes in addNote function, using above same code for api call.
in that code, we used
method: "POST"
in header both content-type and auth-token
Here we send body as like in string mode of json and we send an object containing title and description and tag.
After that we checked if response is ok or not,
Then we call getNotes to show the note is added in browser
We also wrap all this code in useCallback hook.

We also made some changes in editNote function of NoteState.js according to client side.