Editing note by modal. We first add another ref named refClose, to close the modal window after click on update button, and that ref is set to close button as ref={refClose} and in handleClick function of update button, we call that refClose to click like refClose.current.click() by this line the close button is get clicked when we click on update button of modal. Then we add id property to note state of Note.js component as for editing the note we require the id of note which we want to update. Then we called the editNote function of context. And we passed all required arguments to editNote function. And our note is updated.
In editNote of context,
The method is same as addNote,
but difference is we work with id to determine which note we want to update,
so we pass id to fetch fuction also.

In all function of context we directly display the notes awailable in database with help of useCallback fuction and getNotes fuction as UseCallback function's dependency.
As if something is change in notes available in getNotes function then it automatically memorize the setNotes state of getNotes function.
We no need to explicitly show code for front end data chane.