Enhance the view of code. We first set autofocus to add form when page is load and focus to title to make new entry. Then we set the values after adding that note to blank And also if the title and description is blank then add button is set to disable. Also changed input tag of description in add form to textarea tag.
Same we did for the modal form (edit form) and tag change also.

We also show a message if there is no note in database.

Also, we set some style to description part of showing notes in NoteItem.js,
so that we can show the note as it is saved or inputed. ( note is already saved in mongodb database as it is inputed)